+++ sources/Renderer/Direct3D11/Buffer/D3D11IndexBuffer.cpp
@@ -17,7 +17,7 @@ namespace LLGL
 D3D11IndexBuffer::D3D11IndexBuffer(ID3D11Device* device, const BufferDescriptor& desc, const void* initialData) :
     D3D11Buffer { BufferType::Index }
 {
-    CreateResource(device, CD3D11_BUFFER_DESC(desc.size, D3D11_BIND_INDEX_BUFFER), initialData, desc.flags);
+    CreateResource(device, CD3D11_BUFFER_DESC(static_cast<UINT>(desc.size), D3D11_BIND_INDEX_BUFFER), initialData, desc.flags);
     format_ = D3D11Types::Map(desc.indexBuffer.format.GetDataType());
 }
